searchByImage/requirements.txt
tensorflow==2.4.0
fastapi==0.63.0
numpy==1.19.4
Pillow==7.0.0
opencv-python==4.2.0.34
Keras-Preprocessing==1.1.2
